requirements.txt
pandas>=2.2
numpy>=1.26
matplotlib>=3.8
python-dotenv>=1.0
fredapi>=0.5
yfinance>=0.2.40
requests>=2.32
PyYAML>=6.0
